requirements.txt
langchain-community==0.0.10
pypdf==3.15.1
sentence-transformers==2.2.2
chromadb==0.4.18
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.23.2
streamlit==1.28.0
python-dotenv==1.0.0
PyYAML==6.0.1
google-generativeai==0.7.1
# Add other dependencies as needed later 